+++ src/Kubia/Interfaces/BpmRegistratorInterface.php
@@ -4,7 +4,9 @@ namespace Kubia\Interfaces;
 
 use Quancy\Interfaces\CommonInterface;
 
-interface BpmFinalLoggerInterface extends CommonInterface
+interface BpmRegistratorInterface extends CommonInterface
 {
     public function processComplete($data, $headers);
+
+    public function processTimeout($data, $headers);
 }