+++ CHANGELOG.md
@@ -1,17 +1,24 @@
 # Changelog
 
-Logging finalized process instance in Camunda
+Registration various events and logging finalized process instance in Camunda
 
 ## [Unreleased]
 
 ### Planned
 - Nothing
 
+## [1.1] - 2020-01-28
+
+### Added
+
+- Added `processTimeout()` method
+
 ## [1.0] - 2020-01-22
 
 ### Added
 
-- First working version of bpm-final-logger
+- First working version of bpm-registrator
 
-[unreleased]: https://gitlab.com/quancy-core/bpm-final-logger/-/tags/1.0
+[unreleased]: https://gitlab.com/quancy-core/bpm-final-logger/-/tags/1.1
+[1.1]: https://gitlab.com/quancy-core/bpm-final-logger/-/tags/1.1
 [1.0]: https://gitlab.com/quancy-core/bpm-final-logger/-/tags/1.0